Move AssemblyLoadContext to shared partition
